move models inside base_model classes
