+++ src/antares/craft/service/api_services/models/base_model.py
@@ -9,3 +9,9 @@
 # SPDX-License-Identifier: MPL-2.0
 #
 # This file is part of the Antares project.
+from pydantic import BaseModel, ConfigDict
+from pydantic.alias_generators import to_camel
+
+
+class APIBaseModel(BaseModel):
+    model_config = ConfigDict(populate_by_name=True, alias_generator=to_camel, extra="forbid")
