+++ src/antares/craft/service/api_services/models/settings.py
@@ -52,7 +52,7 @@ from antares.craft.model.settings.optimization import (
     UnfeasibleProblemBehavior,
 )
 from antares.craft.model.settings.thematic_trimming import ThematicTrimmingParameters, ThematicTrimmingParametersUpdate
-from antares.craft.service.api_services.models import APIBaseModel
+from antares.craft.service.api_services.models.base_model import APIBaseModel
 from antares.craft.tools.all_optional_meta import all_optional_model
 from pydantic import Field, field_validator
