+++ src/antares/craft/service/api_services/models/thermal.py
@@ -20,7 +20,7 @@ from antares.craft.model.thermal import (
     ThermalClusterPropertiesUpdate,
     ThermalCostGeneration,
 )
-from antares.craft.service.api_services.models import APIBaseModel
+from antares.craft.service.api_services.models.base_model import APIBaseModel
 from antares.craft.tools.all_optional_meta import all_optional_model
 
 ThermalPropertiesType = Union[ThermalClusterProperties, ThermalClusterPropertiesUpdate]
